requirements.txt
beautifulsoup4==4.12.3
soupsieve==2.5
fake-useragent==1.5.1
lxml==5.2.2
requests==2.32.3
selenium==4.23.1
python-dotenv==1.0.1
